requirements.txt
transformers>=4.35.0
torch>=2.0.0
pyyaml>=6.0
rich>=13.0.0
click>=8.1.0
accelerate>=0.20.0
